Migrate `upload-application` command
