+++ packages/base/src/cli.ts
@@ -2,7 +2,8 @@ import {CommandClass} from 'clipanion'
 
 import {DeployTestsCommand} from './commands/synthetics/deploy-tests-command'
 import {RunTestsCommand} from './commands/synthetics/run-tests-command'
+import {UploadApplicationCommand} from './commands/synthetics/upload-application-command'
 
 export const baseCommands: Record<string, CommandClass[]> = {
-  synthetics: [RunTestsCommand, DeployTestsCommand],
+  synthetics: [RunTestsCommand, DeployTestsCommand, UploadApplicationCommand],
 }
